fix: 글 작성 버그 수정
