+++ src/pages/posts/index.tsx
@@ -10,7 +10,19 @@ import { PostReqData } from '@src/types/PostTypes';
 import { FormProvider, SubmitHandler, useForm } from 'react-hook-form';
 
 function PostPage() {
-  const methods = useForm<PostReqData>();
+  const defaultValues: PostReqData = {
+    boardId: 'notice_board',
+    category: 'all',
+    createdAt: '1111-01-01',
+    modifiedAt: '1111-01-01',
+    file: false,
+    hide: false,
+    notice: false,
+    title: '',
+    content: '',
+    photo: 'test',
+  };
+  const methods = useForm<PostReqData>({ defaultValues });
   const { mutate } = usePostArticleMutation();
   const { handleSubmit } = methods;
